Scrolling back from the start of the status screen works.
git-svn-id: http://svn.code.sf.net/p/wp34s/code@1536 9654ca62-1852-461c-aa49-1be32f3bb02d

+++ trunk/keys.c
@@ -1937,7 +1937,7 @@ static int process_confirm(const keycode c) {
 }
 
 static int process_status(const keycode c) {
-	unsigned int n = State2.status - 1;
+	int n = ((int)State2.status) - 1;
 
 	if ( c == K40 ) {
 		if (--n < 0)
